Chore: #31 - 길찾기 함수 utils directory 로 분리
